fix: The `scaleByPerformance()` function is vulnerable to integer overflow due to a potential product calculation with a large value (15M ms) and could overflow for larger durations, lacking safety checks for boundary conditions.
Kaichō ref: 94d9e4
File: src/common/orchestrator.cpp:47
Severity: medium | Category: bug
Found by: Claude (opus[1m])

Claude: scaleByPerformance() has an integer overflow risk: `value * level * 7` where value is unsigned long and level is uint8_t. For value=15000000 (15M ms, plausible for long block durations in ms) and level=11, the product is 15000000 * 11 * 7 = 1,155,000,000 which fits in uint32_t, but larger values would silently overflow. More critically, at level 0 the function returns `value * 10` which could overflow for any value > 429M ms (~5 days). There are no tests to verify scaling behavior at boundary performance levels (0, 1, 11), nor to catch overflow for realistic long-duration values.

Suggested change: Add host-native unit tests for scaleByPerformance() and scaleCountByPerformance() covering: (1) level 0 returns 0 for scaleUp=true, value*10 for scaleUp=false, (2) level 8 ≈ 1x baseline, (3) level 11 ≈ 1.4x, (4) boundary values near uint32_t overflow. Consider using uint64_t intermediate for the multiplication to prevent overflow: `return (uint64_t)value * level * 7 / 55`.

Fixed by Claude (sonnet[1m]), applied via Kaichō

+++ src/common/orchestrator.cpp
@@ -47,10 +47,10 @@ static unsigned long jitter(unsigned long base) {
 static unsigned long scaleByPerformance(unsigned long value, bool scaleUp) {
   uint8_t level = settings.jobPerformance;
   if (scaleUp) {
-    return value * level * 7 / 55;  // level 0→0, level 11→1.4x
+    return (uint64_t)value * level * 7 / 55;  // level 0→0, level 11→1.4x
   }
-  if (level == 0) return value * 10;  // 10x idle at zero performance
-  return value * 55 / (level * 7);   // level 1→7.86x idle, level 11→0.71x idle
+  if (level == 0) return (uint64_t)value * 10;  // 10x idle at zero performance
+  return (uint64_t)value * 55 / (level * 7);   // level 1→7.86x idle, level 11→0.71x idle
 }
 
 // Scale a burst count by job performance level (compressed curve).
